# chore(stdlib): Simplify ip_subnet function code